requirements.txt
python-docx==0.8.11
python-dotenv==1.0.1
openai==1.14.3
pandas==2.2.2
pyyaml==6.0.1  
requests==2.31.0
groq==0.1.1
pytest==7.4.3